Bumped version number to 2.4.0.0. A few tiny changes to make things more type-correct.

+++ Graphics/Rendering/OpenGL/GLU/NURBS.hs
@@ -277,7 +277,7 @@ loadSamplingMatrices nurbsObj =
       (\(mv, proj, (Position x y, Size w h)) -> do
           withMatrixColumnMajor mv $ \mvBuf ->
              withMatrixColumnMajor proj $ \projBuf ->
-                withArray [x, y, w, h] $ \viewportBuf ->
+                withArray [x, y, fromIntegral w, fromIntegral h] $ \viewportBuf ->
                   gluLoadSamplingMatrices nurbsObj mvBuf projBuf viewportBuf
           setAutoLoadMatrix nurbsObj False)
